requirements.txt
flask
firebase-admin
flask-cors
routes
python-dotenv
